Factor out DisplayContextMenu
